requirements.txt
torch==1.10.1+cu113
torchvision==0.11.2+cu113
pytorch-lightning==1.5.9

matplotlib
pandas~=1.4.2
tqdm~=4.64.0
Pillow~=9.1.0
torchmetrics~=0.8.2
albumentations~=1.1.0